Move example screens' testIDs to examples\testIDs
