+++ e2e/calendars.spec.js
@@ -1,4 +1,5 @@
-const testIDs = require('../src/testIDs');
+const {HEADER_MONTH_NAME, CHANGE_MONTH_LEFT_ARROW, CHANGE_MONTH_RIGHT_ARROW} = require('../src/testIDs');
+const testIDs = require('../example/src/testIDs');
 
 describe('Calendars', () => {
   const FIRST_CALENDAR = testIDs.calendars.FIRST;
@@ -14,17 +15,17 @@ describe('Calendars', () => {
   });
 
   it('should move to previous month', async () => {
-    await expect(element(by.id(`${testIDs.HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('February 2020');
+    await expect(element(by.id(`${HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('February 2020');
 
-    await element(by.id(`${testIDs.CHANGE_MONTH_LEFT_ARROW}-${FIRST_CALENDAR}`)).tap();
-    await expect(element(by.id(`${testIDs.HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('January 2020');
+    await element(by.id(`${CHANGE_MONTH_LEFT_ARROW}-${FIRST_CALENDAR}`)).tap();
+    await expect(element(by.id(`${HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('January 2020');
   });
 
   it('should move to next month twice', async () => {
-    await expect(element(by.id(`${testIDs.HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('February 2020');
+    await expect(element(by.id(`${HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('February 2020');
 
-    await element(by.id(`${testIDs.CHANGE_MONTH_RIGHT_ARROW}-${FIRST_CALENDAR}`)).tap();
-    await element(by.id(`${testIDs.CHANGE_MONTH_RIGHT_ARROW}-${FIRST_CALENDAR}`)).tap();
-    await expect(element(by.id(`${testIDs.HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('April 2020');
+    await element(by.id(`${CHANGE_MONTH_RIGHT_ARROW}-${FIRST_CALENDAR}`)).tap();
+    await element(by.id(`${CHANGE_MONTH_RIGHT_ARROW}-${FIRST_CALENDAR}`)).tap();
+    await expect(element(by.id(`${HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('April 2020');
   });
 });
