+++ e2e/horizontalList.spec.js
@@ -1,4 +1,5 @@
-const testIDs = require('../src/testIDs');
+const {HEADER_MONTH_NAME} = require('../src/testIDs');
+const testIDs = require('../example/src/testIDs');
 
 describe('Horizontal Calendar List', () => {
   const FIRST_CALENDAR = `${testIDs.horizontalList.CONTAINER}_1273968000000`;
@@ -11,13 +12,13 @@ describe('Horizontal Calendar List', () => {
 
   it('should scroll calendars to the top', async () => {
     await element(by.id(testIDs.horizontalList.CONTAINER)).scrollTo('left');
-    await expect(element(by.id(`${testIDs.HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toBeVisible();
-    await expect(element(by.id(`${testIDs.HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('May 2010');
+    await expect(element(by.id(`${HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toBeVisible();
+    await expect(element(by.id(`${HEADER_MONTH_NAME}-${FIRST_CALENDAR}`))).toHaveText('May 2010');
   });
 
   it('should scroll calendars to the bottom', async () => {
     await element(by.id(testIDs.horizontalList.CONTAINER)).scrollTo('right');
-    await expect(element(by.id(`${testIDs.HEADER_MONTH_NAME}-${LAST_CALENDAR}`))).toBeVisible();
-    await expect(element(by.id(`${testIDs.HEADER_MONTH_NAME}-${LAST_CALENDAR}`))).toHaveText('May 2014');
+    await expect(element(by.id(`${HEADER_MONTH_NAME}-${LAST_CALENDAR}`))).toBeVisible();
+    await expect(element(by.id(`${HEADER_MONTH_NAME}-${LAST_CALENDAR}`))).toHaveText('May 2014');
   });
 });
