+++ example/src/screens/agenda.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import {Alert, StyleSheet, Text, View, TouchableOpacity} from 'react-native';
 import {Agenda} from 'react-native-calendars';
 
-const testIDs = require('../../../src/testIDs');
+const testIDs = require('../testIDs');
 
 
 export default class AgendaScreen extends Component {
