+++ example/src/screens/horizontalCalendarList.js
@@ -1,8 +1,7 @@
 import React, {Component} from 'react';
 import {CalendarList} from 'react-native-calendars';
 
-const testIDs = require('../../../src/testIDs');
-
+const testIDs = require('../testIDs');
 
 export default class HorizontalCalendarList extends Component {
